requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
gunicorn==23.0.0
idna==3.11
pillow==12.1.1
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
